# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Blog",
        "description": "Documentación de la API (Laravel 11 + Passport).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor base"
        }
    ],
    "paths": {
...
        "/api/card-translations": {
            "get": {
                "tags": [
                    "Card Translations"
                ],
                "summary": "Muestra una lista de todas las traducciones de tarjetas.",
                "operationId": "getCardTranslationsIndex",
                "responses": {
                    "200": {
                        "description": "Lista de traducciones obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CardTranslationResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado."
                    },
                    "403": {
                        "description": "Acción no autorizada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Card Translations"
                ],
                "summary": "Crea una nueva traducción para una tarjeta. Genera audio TTS si 'audio_file' no se proporciona.",
                "operationId": "createCardTranslation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "card_id_translation",
                                    "language_code",
                                    "key_phrase"
                                ],
                                "properties": {
                                    "card_id_translation": {
                                        "description": "ID de la tarjeta a traducir.",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "language_code": {
                                        "description": "Código del idioma (e.g., es, en).",
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "key_phrase": {
                                        "description": "Frase clave a traducir.",
                                        "type": "string",
                                        "example": "Hello, world"
                                    },
                                    "audio_file": {
                                        "description": "Archivo de audio MP3 opcional. Si no se provee, se usa TTS.",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Traducción creada y audio generado (o subido) exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardTranslationResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado."
                    },
                    "403": {
                        "description": "Acción no autorizada (solo terapeutas/administradores)."
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/card-translations/{id}": {
            "get": {
                "tags": [
                    "Card Translations"
                ],
                "summary": "Muestra una traducción específica por ID.",
                "operationId": "getCardTranslationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la traducción.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Traducción obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardTranslationResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado."
                    },
                    "403": {
                        "description": "Acción no autorizada."
                    },
                    "404": {
                        "description": "Traducción no encontrada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Card Translations"
                ],
                "summary": "Actualiza una traducción existente por ID.",
                "description": "Se debe usar POST con campo _method=PUT/PATCH en multipart/form-data. Genera audio TTS si 'audio_file' no se proporciona y se cambia la 'key_phrase'.",
                "operationId": "updateCardTranslation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la traducción a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Método HTTP para simular PUT/PATCH.",
                                        "type": "string",
                                        "enum": [
                                            "PUT",
                                            "PATCH"
                                        ],
                                        "example": "PATCH"
                                    },
                                    "language_code": {
                                        "description": "Código del idioma (e.g., es, en).",
                                        "type": "string",
                                        "example": "fr",
                                        "nullable": true
                                    },
                                    "key_phrase": {
                                        "description": "Nueva frase clave.",
                                        "type": "string",
                                        "example": "Bonjour le monde",
                                        "nullable": true
                                    },
                                    "audio_file": {
                                        "description": "Nuevo archivo de audio MP3 opcional. Si no se provee, se usa el audio existente o se genera TTS si la frase cambia.",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Traducción actualizada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardTranslationResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado."
                    },
                    "403": {
                        "description": "Acción no autorizada."
                    },
                    "404": {
                        "description": "Traducción no encontrada."
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Card Translations"
                ],
                "summary": "Elimina una traducción específica por ID.",
                "operationId": "deleteCardTranslation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la traducción a eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Traducción eliminada exitosamente (No Content)."
                    },
                    "401": {
                        "description": "No autenticado."
                    },
                    "403": {
                        "description": "Acción no autorizada."
                    },
                    "404": {
                        "description": "Traducción no encontrada."
                    },
                    "500": {
                        "description": "Error interno del servidor al intentar eliminar."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CardTranslationResource": {
                "title": "Card Translation Resource",
                "description": "Estructura de datos para la traducción de una tarjeta, incluyendo la ruta del archivo de audio TTS/subido.",
                "properties": {
                    "translationId": {
                        "description": "ID único de la traducción.",
                        "type": "integer",
                        "example": 1
                    },
                    "cardId": {
                        "description": "ID de la tarjeta a la que pertenece esta traducción.",
                        "type": "integer",
                        "example": 10
                    },
                    "languageCode": {
                        "description": "Código del idioma de la traducción (e.g., 'en', 'es').",
                        "type": "string",
                        "example": "en"
                    },
                    "keyPhrase": {
                        "description": "La frase clave traducida (el texto que se reproduce).",
                        "type": "string",
                        "example": "Hello, world"
                    },
                    "audioPath": {
                        "description": "Ruta del archivo de audio generado por TTS o subido.",
                        "type": "string",
                        "example": "user_1/audio/translation_abc123.mp3"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}